fix: Rename webui/lib to webui/shared to resolve import collision
The webui/lib directory was shadowing the firmware lib/ directory,
causing "ModuleNotFoundError: No module named 'lib.gps_exif_lib'"
when the webui backend tried to import GPS EXIF functions.

Renaming to webui/shared eliminates the naming collision and allows
clean imports without sys.path manipulation.
