+++ Firmware/CLAUDE.md
@@ -440,8 +440,8 @@ def ws_connect():
 ### GPS Coordinate Conversion
 
 ```python
-# CORRECT - Use shared utilities from webui.lib
-from webui.lib.gps_coordinates import decimal_to_dms, dms_to_decimal
+# CORRECT - Use shared utilities from webui.shared
+from webui.shared.gps_coordinates import decimal_to_dms, dms_to_decimal
 
 # Convert decimal to DMS for EXIF embedding
 lat_dms = decimal_to_dms(37.7749, is_latitude=True)
@@ -476,7 +476,7 @@ const lonDisplay = formatCoordinateDisplay(-122.4194, false);
 - `install_mothbox.sh`: Installation script with Pi detection and firmware selection
 - `gps_exif_tagger.py`: Main GPS EXIF embedding tool (CLI and watch mode)
 - `lib/gps_exif_lib.py`: GPS EXIF library (coordinate conversion, EXIF embedding, verification)
-- `webui/lib/gps_coordinates.py`: **GPS coordinate utilities** (decimal ↔ DMS conversion, validation, formatting) - webui-shared library
+- `webui/shared/gps_coordinates.py`: **GPS coordinate utilities** (decimal ↔ DMS conversion, validation, formatting) - webui-shared library
 - `webui/frontend/src/utils/gpsCoordinates.ts`: **GPS coordinate utilities (TypeScript)** (identical behavior to Python)
 - `webui/backend/app.py`: Flask app initialization, CSRF, CORS, SocketIO setup
 - `webui/backend/liveview_stream.py`: Camera streaming engine (2500+ lines)
