+++ Firmware/webui/shared/gps_coordinates.py
